requirements.txt
pydantic>=2,<3
pyyaml>=6,<7
typer>=0.9,<1
rich>=13,<15
httpx>=0.27,<1
duckduckgo-search>=6,<7
openai-whisper>=20231117
soundfile>=0.12,<1
numpy>=1,<3

# Optional — install based on your chosen providers:
# kokoro>=0.9             # TTS: Kokoro
# TTS>=0.22              # TTS: Coqui
# diffusers>=0.30        # Visuals: FLUX.1
# torch>=2               # Visuals: FLUX.1
# Pillow>=10,<11         # Visuals
# anthropic>=0.40,<1     # LLM: Claude
